Move app inside components folder
